requirements.txt
altgraph
colorama
customtkinter
darkdetect
packaging
pefile
pillow
pyinstaller